Refactoring, move old outlier filter into new file
