+++ src/modules/src/kalman_core/Kbuild
@@ -11,4 +11,5 @@ obj-y += mm_tdoa_robust.o
 obj-y += mm_tof.o
 obj-y += mm_yaw_error.o
 obj-y += outlierFilterTdoa.o
+obj-y += outlierFilterTdoaSteps.o
 obj-y += outlierFilterLighthouse.o
